refactor proposal version routes
